+++ bansheed/src/config.rs
@@ -18,7 +18,7 @@ impl Default for DaemonConfig {
     }
 }
 
-#[derive(Deserialize, Serialize, Debug, Clone, Copy)]
+#[derive(Deserialize, Serialize, Debug, Clone, Copy, PartialEq, Eq)]
 #[serde(rename_all = "lowercase")]
 pub enum HotkeyMode {
     Hold,